Unsequenced argument evaluation was real and there were TWO. Under MSVC's order
every external pass registered under the bare prefix 'External:' -- one key for
all of them, each registration evicting the last, and unregistration never
matching.

WHY Г24's census missed it, which is the most useful answer here: that gate opens
ONE file by absolute path and counts calls to ONE function named in a constant.
Either narrowness is fatal. Its reach never opened SceneRenderer.cpp, and its
subject is a NAME -- so even aimed at the whole tree it would find nothing, because
here the consuming side effect is spelled std::move and the reading one is a plain
member access. The new gate scans six source roots for the SHAPE.

memcmp: padding is the right argument for one site and the wrong one for
CloudTypeShape, which is all floats -- and no float has a unique object
representation, so -0.0f and 0.0f were one authored value with two bit patterns
forcing a re-bake of an unchanged sky.

Ten of fourteen rounding sites clamped first and were already correct, which is
what made the form dangerous: the four that did not looked identical.

# Conflicts:
#	.clang-tidy
#	Desert/Desert/Source/Engine/Generated/Reflection.gen.cpp
#	Desert/Desert/Source/Engine/Graphic/Render/RenderCommand.hpp

+++ Desert/Desert/Source/Engine/Assets/CloudNoiseVolume.cpp
@@ -215,7 +215,7 @@ namespace Desert::Assets
 
         for ( uint32_t channel = 0; channel < 4u; ++channel )
         {
-            const uint32_t stored = ReadU32( at + 20 + channel * 4u );
+            const uint32_t stored = ReadU32( at + 20u + static_cast<size_t>( channel ) * 4u );
             if ( stored != channel )
                 return Common::MakeFormattedError<CloudNoiseVolumeData>(
                      "channel {} declares meaning {}, but this build reads volumes whose channels are in the "
